+++ lib/oneview-sdk/resource/api300/synergy/volume_attachment.rb
@@ -14,6 +14,7 @@ require_relative '../../api200/volume_attachment'
 module OneviewSDK
   module API300
     module Synergy
+      # Storage volume attachment resource implementation for API300 Synergy
       class VolumeAttachment < OneviewSDK::API200::VolumeAttachment
       end
     end