+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  1. Added full support to OneView Rest API version 300 for the hardware variants C7000 and Synergy to the already existing features:
    - Logical Interconnect
    - Uplink Set
+   - Volume attachment
 
 ### Features supported
 - Connection template
@@ -33,6 +34,7 @@
 - Storage systems
 - Switches
 - Volume
+- Volume attachment
 - Volume template
 - Drive Enclosures
 - Interconnect Link Topology